feat: 위시 상품 삭제 API 예외 추가
feat: 위시 상품 삭제 API 예외 추가

+++ src/main/java/gift/wish/service/WishServiceImpl.java
@@ -6,6 +6,7 @@ import gift.wish.dto.WishCreateResponseDto;
 import gift.wish.dto.WishGetResponseDto;
 import gift.wish.dto.WishPageResponseDto;
 import gift.wish.entity.Wish;
+import gift.wish.exception.ForbiddenException;
 import gift.wish.repository.WishRepository;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -22,6 +23,7 @@ public class WishServiceImpl implements WishService {
 
     @Override
     public WishCreateResponseDto addWish(Member member, WishCreateRequestDto wishCreateRequestDto) {
+        // TODO: 이미 추가한 상품인지 확인하기(WishRepository.existsByMemberAndProduct) 실패 시 예외 처리(이미 존재하는 위시)
 
         Wish wish = new Wish(member.getMemberId(), wishCreateRequestDto.productId());
         wishRepository.addWish(wish);
@@ -53,7 +55,15 @@ public class WishServiceImpl implements WishService {
     }
 
     @Override
-    public void deleteWish(Long wishId) {
+    public void deleteWish(Member member, Long wishId) {
+        // TODO: 위시 ID 검색(WishRepository.findByWishId), 실패 시 예외 처리(존재하지 않은 위시)
+        Wish wish = wishRepository.findByWishId(wishId);
+
+        // TODO: 위시 데이터에 존재하는 회원 ID와 제공한 회원 ID 비교, 실패 시 예외 처리(다른 사용자)
+        if (!member.getMemberId().equals(wish.getMemberId())) {
+            throw new ForbiddenException("다른 사용자의 위시리스트에 접근할 수 업습니다.");
+        }
+
         wishRepository.deleteWish(wishId);
     }
